update python tests
