+++ raphtory-graphql/schema.graphql
@@ -5644,4 +5644,3 @@ schema {
 	query: QueryRoot
 	mutation: MutRoot
 }
-
